Move YouTubeActivity into chapter package
